[cli] Make --default-async-timeout available from cli
Because I can allow an overrided event loop, I need to load the plugin
dynamically in pytest_configure

Unfortunately this is not called before pytest looks at cli options and
so it's not available at that time.

My compromise is to load the options separately to the plugin itself as
part of loading the plugin

This also means that I need pytest to be available at the time of
installing the package because setup.py finds alt_pytest_asyncio.VERSION
so I have moved that into it's own file and get it from there instead
and then put it in alt_pytest_asyncio.__init__ so you can still
programmatically get that number the same way.